requirements.txt
openai>=2.8.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pytest>=8.0.0
git+https://github.com/openai/whisper.git
